+++ android/chat-compose/build.gradle.kts
@@ -61,6 +61,7 @@ android {
 
 dependencies {
     implementation("com.joinself:mobile-sdk:1.0.0-SNAPSHOT")
+    implementation(project(":common"))
 //    implementation(project(":self-android-sdk"))
 
     implementation("com.jakewharton.timber:timber:${Config.Version.timberVersion}")
