+++ android/chat/build.gradle.kts
@@ -66,6 +66,7 @@ android {
 
 dependencies {
     implementation("com.joinself:mobile-sdk:1.0.0-SNAPSHOT")
+    implementation(project(":common"))
 //    implementation(project(":self-android-sdk"))
 
     implementation("com.jakewharton.timber:timber:${Config.Version.timberVersion}")
