+++ android/chat/src/main/java/com/joinself/sdk/sample/MainActivity.kt
@@ -18,8 +18,4 @@ class MainActivity : AppCompatActivity() {
         setContentView(binding.root)
 
     }
-
-    override fun onNewIntent(intent: Intent?) {
-        super.onNewIntent(intent)
-    }
 }
