--- android/sign-in/src/main/java/com/joinself/sdk/sample/Constants.kt
@@ -1,39 +1,0 @@
-package com.joinself.sdk.sample
-
-class Constants {
-    companion object {
-        const val CLAIM_KEY_EMAIL = "email_address"
-        const val CLAIM_KEY_PHONE = "phone_number"
-        const val CLAIM_KEY_PHOTO = "photo"
-        const val CLAIM_KEY_NAME = "display_name"
-        const val CLAIM_KEY_GIVEN_NAME = "given_names"
-        const val CLAIM_KEY_SURNAME = "surname"
-        const val CLAIM_KEY_SEX = "sex"
-        const val CLAIM_KEY_NUMBER = "document_number"
-        const val CLAIM_KEY_ADDRESS = "address"
-        const val CLAIM_KEY_NATIONALITY = "nationality"
-        const val CLAIM_KEY_DOB = "date_of_birth"
-        const val CLAIM_KEY_POB = "place_of_birth"
-        const val CLAIM_KEY_DATE_OF_EXPIRATION = "date_of_expiration"
-        const val CLAIM_KEY_DATE_OF_ISSUANCE = "date_of_issuance"
-        const val CLAIM_KEY_ISSUE_AUTHORITY = "issuing_authority"
-        const val CLAIM_KEY_VALID_FROM = "valid_from"
-        const val CLAIM_KEY_VALID_TO = "valid_to"
-        const val CLAIM_KEY_CATEGORIES = "categories"
-        const val CLAIM_KEY_SORT_CODE = "sort_code"
-        const val CLAIM_KEY_SELFIE_HASH = "selfie_hash"
-        const val CLAIM_KEY_IMAGE_HASH = "image_hash"
-        const val CLAIM_KEY_COUNTRY_OF_ISSUANCE = "country_of_issuance"
-        const val CLAIM_KEY_MRZ= "document_mrz"
-        const val CLAIM_KEY_SELFIE_VERIFICATION = "selfie_verification"
-
-        const val FIELD_ACCOUNT_ID = "account_id"
-        const val FIELD_NICKNAME = "nickname"
-
-        const val FIELD_VERIFIED_ON = "verified_on"
-        const val FIELD_VERIFIED_BY = "verified_by"
-        const val FIELD_LAST_VERIFIED_ON = "last_verified_on"
-        const val FIELD_LAST_VERIFIED_BY = "last_verified_by"
-        const val FIELD_ADDED_ON = "added_on"
-    }
-}
